+++ e2e-tests/src/finalization.rs
@@ -1,0 +1,9 @@
+use common::create_connection;
+
+use crate::config::Config;
+use crate::utils::waiting::wait_for_finalized_block;
+
+pub fn test_finalization(config: Config) -> anyhow::Result<u32> {
+    let connection = create_connection(config.node);
+    wait_for_finalized_block(&connection, 1)
+}
