+++ e2e-tests/src/treasury.rs
@@ -7,17 +7,17 @@ use common::create_connection;
 use frame_support::PalletId;
 use log::info;
 use sp_core::Pair;
-use sp_runtime::MultiAddress;
 use sp_runtime::traits::AccountIdConversion;
-use substrate_api_client::{AccountId, UncheckedExtrinsicV4};
+use sp_runtime::MultiAddress;
 use substrate_api_client::sp_runtime::AccountId32;
+use substrate_api_client::{AccountId, UncheckedExtrinsicV4};
 
 use crate::config::Config;
-use crate::utils::{setup_for_transfer, transfer, };
 use crate::utils::accounts::{accounts, get_free_balance, get_sudo};
 use crate::utils::fee::get_tx_fee_info;
+use crate::utils::transfer::{setup_for_transfer, transfer};
 use crate::utils::types::Connection;
-use crate::waiting::wait_for_event;
+use crate::utils::waiting::wait_for_event;
 
 pub fn test_channeling_fee(config: Config) -> anyhow::Result<()> {
     let (connection, _, to) = setup_for_transfer(config);
