+++ e2e-tests/src/utils/extrinsic.rs
@@ -1,0 +1,27 @@
+#[macro_export]
+macro_rules! send_extrinsic {
+	($connection: expr,
+	$module: expr,
+	$call: expr,
+    $hash_log: expr
+	$(, $args: expr) *) => {
+		{
+            use substrate_api_client::{compose_extrinsic, UncheckedExtrinsicV4, XtStatus};
+
+            let tx: UncheckedExtrinsicV4<_> = compose_extrinsic!(
+                $connection,
+                $module,
+                $call
+                $(, ($args)) *
+            );
+
+            let tx_hash = $connection
+                .send_extrinsic(tx.hex_encode(), XtStatus::Finalized)
+                .unwrap()
+                .expect("Could not get tx hash");
+            $hash_log(tx_hash);
+
+            tx
+		}
+    };
+}
